Code Refactoring Issue Fixed. Deleted old files and fixed imports
