+++ casino_games/blackjack.py
@@ -1,9 +1,8 @@
 # BLACKJACK
 import random
-import screenspace as ss
-from screenspace import Terminal
-from style import graphics as g
-
+from utils.utils import screenspace as ss
+from utils.utils import Terminal
+from utils.utils import g # g is graphics
 game_title = "🃑 Blackjack"
 
             # 0         1         2      3      4
