+++ casino_games/coin_flip.py
@@ -1,9 +1,9 @@
 # COIN FLIP
 import random
 from time import sleep
-import screenspace as ss
-from screenspace import Terminal
-from style import graphics as g
+from utils.utils import screenspace as ss
+from utils.utils import Terminal
+from utils.utils import g # g is graphics
 
 game_title = "⛁ Coin Flip"
 header = "─" * ((75 - len(game_title)) // 2) + game_title + "─" * ((75 - len(game_title)) // 2)
