+++ modules_directory/loan.py
@@ -1,12 +1,12 @@
 import keyboard
 import os
-import style
+from utils import screenspace
 import time
 import re
 import networking as net
 from socket import socket
 
-g = style.get_graphics()
+g = screenspace.get_graphics()
 
 
 class LoanScreen:
@@ -85,24 +85,24 @@ class LoanScreen:
         Builds dipslay... that's about it
         
         """
-        retval = style.set_cursor_str(0, 0) + g.get('loan')
+        retval = screenspace.set_cursor_str(0, 0) + g.get('loan')
         
         if self.loanScreen and not self.gettingLoan:
-            retval += style.set_cursor_str(3,
-                                           3) + "== Welcome to the office of J.G. Hopworth, esteemed financial lender =="
-            retval += style.set_cursor_str(30, 5) + "Would you like to take out a loan? (y/n)"
+            retval += screenspace.set_cursor_str(3,
+                                                 3) + "== Welcome to the office of J.G. Hopworth, esteemed financial lender =="
+            retval += screenspace.set_cursor_str(30, 5) + "Would you like to take out a loan? (y/n)"
             
             
         
         elif self.loanScreen and self.gettingLoan:
-            retval += style.set_cursor_str(32, 3) + "You may choose a high or low interest loan."
-            retval += style.set_cursor_str(33, 5) + "High interest (h) - Up to $2000"
-            retval += style.set_cursor_str(33, 6) + "Low interest (l) - Up to $500"
-            retval += style.set_cursor_str(33, 8) + "Press (q) to cancel"
+            retval += screenspace.set_cursor_str(32, 3) + "You may choose a high or low interest loan."
+            retval += screenspace.set_cursor_str(33, 5) + "High interest (h) - Up to $2000"
+            retval += screenspace.set_cursor_str(33, 6) + "Low interest (l) - Up to $500"
+            retval += screenspace.set_cursor_str(33, 8) + "Press (q) to cancel"
             
             if self.low_or_high is not None:
                 loan_type = "high" if self.low_or_high else "low"
-                retval += style.set_cursor_str(33, 10) + f"Selected: {loan_type} interest loan"
+                retval += screenspace.set_cursor_str(33, 10) + f"Selected: {loan_type} interest loan"
         
         return retval
