+++ monopoly_directory/monopoly.py
@@ -7,8 +7,8 @@ from monopoly_directory.properties import Property
 from monopoly_directory.cards import Cards
 from monopoly_directory.board import Board
 from monopoly_directory.player_class import MonopolyPlayer
-from screenspace import calibrate_screen, make_fullscreen, clear_screen
-from style import MYCOLORS as COLORS, set_cursor_str, graphics as g
+from utils.utils import calibrate_screen, make_fullscreen, clear_screen
+from utils.utils import MYCOLORS as COLORS, set_cursor_str, g # graphics is g
 
 mode = "normal"
 output = ""
