+++ utils/utils.py
@@ -1,3 +1,32 @@
+# Loading animation function
+import itertools
+import sys
+import time
+loading = True
+def loading_animation():
+    for frame in itertools.cycle(['|', '/', '-', '\\']):
+        if not loading:
+            break
+        sys.stdout.write(f'\rLoading {frame}')
+        sys.stdout.flush()
+        time.sleep(0.1)
+    sys.stdout.write('\rLoading complete!     \n')
+
+import socket
+import modules_directory.inventory as inv
+class Client:
+    def __init__(self, socket: socket.socket, id: int, name: str, inventory_object: inv.Inventory):
+        self.socket = socket
+        self.id = id
+        self.name = name
+        self.inventory = inventory_object
+        self.can_roll = True
+        self.num_rolls = 0
+        self.terminal_statuses = ["ACTIVE", "ACTIVE", "ACTIVE", "ACTIVE"]
+        self.trades = [{"name":"", "properties":[]}, {"name":"", "properties":[]}, {"name":"", "properties":[]}] # List of trades for this player
+        self.PlayerObject = None # Player object for this client
+
+
 # Written by @https://github.com/SerpentBTW
 
 """
@@ -6,6 +35,7 @@ gameplay. They are only to be called during the setup of the game, where
 printing directly using print() is not an issue. Do not use in any main()
 functions or modules that are called during gameplay.
 """
+
 
 def validate_name(name) -> bool:
     """
@@ -21,7 +51,10 @@ def validate_name(name) -> bool:
 
     return True
 
+
 import subprocess
+
+
 def is_port_unused(port: int) -> bool:
     """
     Check if port is being used in any capacity. If the port inputted is, the program will return False.
@@ -32,11 +65,12 @@ def is_port_unused(port: int) -> bool:
         return False
     return True
 
+
 def validate_port(port: str):
     """
     Validate a port. The port must be a number between 1024 and 65535.
     """
-    
+
     if not port.isdigit():
         print(f"Port {port} is not a number.")
         return False
@@ -47,6 +81,7 @@ def validate_port(port: str):
         return False
 
     return True
+
 
 def validate_address(address: str) -> bool:
     """
